+++ sdk/search/Azure.Search.Documents/src/Indexes/Models/WebApiSkill.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System.Collections.Generic;
-using Azure.Core;
 
 namespace Azure.Search.Documents.Indexes.Models
 {
@@ -10,9 +9,5 @@ namespace Azure.Search.Documents.Indexes.Models
     {
         /// <summary> The headers required to make the http request. </summary>
         public IDictionary<string, string> HttpHeaders { get; }
-
-        /// <summary> The URI of the Web API providing the vectorizer. </summary>
-        [CodeGenMember("Url")]
-        public string Uri { get; set; }
     }
 }